\name{sjedrp-internal}
\alias{binit2}
\alias{drpcorrections}
\alias{drpeffrad}
\alias{drppackingfactor}
\alias{drpreliability}
\alias{binit2.check}
\alias{drp.makemac}
\alias{C_cross_corr_r}
\alias{C_drp_bin_it_r}
\alias{autocorr}
\alias{crosscorr}
\alias{plot.sjecorr}
\title{Internal sjedrp functions}
\description{
  Internal sjedrp functions.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
