\name{sjedrp-internal}
\alias{binit2}
\alias{drpcorrections}
\alias{drpeffrad}
\alias{drppackingfactor}
\alias{drpreliability}
\alias{binit2.check}
\alias{drp.makemac}
\title{Internal sjedrp functions}
\description{
  Internal sjedrp functions.
}
\usage{
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
